rename modules to make things more consistens and logical, #45, #46
