+++ bacpipe/embedding_generation_pipelines/dimensionality_reduction/t_sne.py
@@ -1,4 +1,4 @@
-from ..utils import ModelBaseClass
+from ..model_utils import ModelBaseClass
 from sklearn.manifold import TSNE
