+++ bacpipe/embedding_generation_pipelines/feature_extractors/audiomae.py
@@ -21,7 +21,7 @@ from timm.models.layers import to_2tuple
 
 import bacpipe.model_specific_utils.audiomae.models_vit as models_vit
 from bacpipe.model_specific_utils.audiomae.dataset import AudiosetDataset
-from ..utils import ModelBaseClass
+from ..model_utils import ModelBaseClass
 
 BATCH_SIZE = 8  # important to lower this if run on laptop cpu
