+++ bacpipe/embedding_generation_pipelines/feature_extractors/avesecho_passt.py
@@ -1,5 +1,5 @@
 from hear21passt.base import get_basic_model, get_model_passt
-from ..utils import ModelBaseClass
+from ..model_utils import ModelBaseClass
 import torch.nn as nn
 import torchaudio
 import torch
