+++ bacpipe/embedding_generation_pipelines/feature_extractors/insect459.py
@@ -3,7 +3,7 @@ from types import SimpleNamespace
 import yaml
 import torch
 from .insect66 import SpectrogramCNN
-from ..utils import ModelBaseClass
+from ..model_utils import ModelBaseClass
 
 SAMPLE_RATE = 44100
 LENGTH_IN_SAMPLES = int(5.5 * SAMPLE_RATE)
