+++ bacpipe/embedding_generation_pipelines/feature_extractors/insect66.py
@@ -5,7 +5,7 @@ import torch
 import timm
 import torch.nn as nn
 import torchaudio as ta
-from ..utils import ModelBaseClass
+from ..model_utils import ModelBaseClass
 
 SAMPLE_RATE = 44100
 LENGTH_IN_SAMPLES = int(5.5 * SAMPLE_RATE)
